+++ R/kinematics.R
@@ -7,3 +7,13 @@
 get_velocity <- function(signal) {
   return(c(NA, diff(signal)))
 }
+
+#' Get acceleration of displacement
+#'
+#' @param signal The signal to get the acceleration of.
+#'
+#' @return A vector with the second derivative of the signal.
+#' @export
+get_acceleration <- function(signal) {
+  return(c(NA, NA, diff(signal, differences = 2)))
+}