+++ NEWS.md
@@ -19,7 +19,7 @@
 
 * `resample_signal()` to up/down-sample a signal using interpolation.
 
-* `get_velocity()` to calculate the velocity of displacement.
+* `get_velocity()` and `get_acceleration()` to calculate the velocity and acceleration (first and second derivative) of displacement.